% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities-mapping.R
\name{.stringWrap}
\alias{.stringWrap}
\title{.stringWrap}
\usage{
.stringWrap(label, width = tlfEnv$maxCharacterWidth)
}
\arguments{
\item{label}{A character to be wrapped}

\item{width}{An integer width for line break split algorithm}
}
\description{
Wrap character string by applying line breaks when appropriate
}
\keyword{internal}
